+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0] - 2021-02-20
+
+### Added
+- Update extension version functionality.
+- Add Github Actions workflow to run tests in the version manager module.
+
 ## [0.0.0] - 2021-02-14
 
 ### Added